Revert to standard error handling; upgrade pytest.
This reverts commit e4dfae40860f4fdc224c4c8065d77c45f9c36bb8, which
tries to fix #74 by importing HTTPError from `python_http_client`. That
change isn't necessary, since the error in #74 was being caused by a bug
in `sendgrid-python` which resulted in the wrong version of
`python_http_client` being installed:
https://github.com/sendgrid/sendgrid-python/issues/321

Since the bug is fixed in `sendgrid-python` 4.2.1, but this library
still uses `sendgrid-python` 3.5+, I've added additional installation
constraints on `python_http_client` so that it should never install a
version with custom error handling (2.3.0 or above). This constraint
should be removed when we upgrade to `sendgrid-python` 4.

This commit also adds a minimum version to pytest. pytest-django
requires pytest>=3.6. To prevent test failures, we need to ensure that
an up-to-date version of pytest is installed.